DropFly-PROJECTS/dropfly-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "dropfly-agent"
version = "0.1.0"
description = "Autonomous builder system powered by specialized brains and parallel multi-agent execution"
requires-python = ">=3.12"
dependencies = [
    # Core
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",

    # LLM SDKs (all optional — install what you need)
    "anthropic>=0.40.0",

    # Database
    "supabase>=2.0.0",

    # API server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "PyJWT>=2.8.0",

    # Tools
    "playwright>=1.48.0",
    "gitpython>=3.1.0",
    "docker>=7.0.0",

    # Async
    "anyio>=4.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.50.0"]
google = ["google-generativeai>=0.8.0"]
mistral = ["mistralai>=1.0.0"]
ollama = ["ollama>=0.3.0"]
all-providers = [
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    "mistralai>=1.0.0",
    "ollama>=0.3.0",
]
communication = [
    "twilio>=9.0.0",
    "elevenlabs>=1.0.0",
]
search = [
    "tavily-python>=0.5.0",
]
aws = [
    "boto3>=1.35.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
dropfly = "src.cli:main"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
